Fix for [#HER-1097] ARCWriter cannot handle records larger than 2 GB * (10 classes)     change relevant fields, method parameters, settings, and constants to be long/Long rather than int/Integer

+++ heritrix/src/java/org/archive/io/WriterPoolSettings.java
@@ -32,7 +32,7 @@ import java.util.List;
  * @version $Date$, $Revision$
  */
 public interface WriterPoolSettings {
-    public int getMaxSize();
+    public long getMaxSize();
     public String getPrefix();
     public String getSuffix(); 
     public List<File> getOutputDirs();
